EBP-2658: experimental Windows amd64 support via shared CCSMP lib
Add opt-in experimental Windows (x86_64) support, gated behind the build
tag `windows_shared_experimental`. Default builds on all platforms are
unchanged; a default (untagged) Windows build fails fast with guidance.

- Link dynamically against the MSVC (VS2015) shared library via its
  import lib (internal/ccsmp/lib/windows_amd64/libsolclient.lib);
  libsolclient.dll is required at runtime.
- includes_windows_amd64_common.go: CFLAGS incl. forced winsock2.h.
- includes_windows_amd64.go: tag-gated LDFLAGS (-l:libsolclient.lib).
- includes_windows_amd64_unsupported.go: #error guard for default builds.
- dummy.go: vendor the windows_amd64 lib dir.
- EBP-2653: fix LLP64 integer sizing -- use C.size_t (not C.ulong) for
  size_t parameters in the message dump and message tracing paths.

OpenSSL (libcrypto-3.dll/libssl-3.dll) is runtime-loaded for TLS only and
is not redistributed by this module (same model as darwin).

+++ internal/ccsmp/ccsmp_message.go
@@ -661,7 +661,7 @@ func SolClientMessageDump(messageP SolClientMessagePt) string {
 
 	// Truncate the message after 10,000 characters, SOL-62945
 	// removed the dynamic calculation of buffer size as defaultMsgDumpBufferSize{1000} + (payloadSize * msgDumpMultiplier{5})
-	bufferSize := C.ulong(maxDumpSize)
+	bufferSize := C.size_t(maxDumpSize)
 	buffer := (*C.char)(C.malloc(bufferSize))
 	defer C.free(unsafe.Pointer(buffer))
